Combine two files -loaded in the editor- into one
