+++ src/editor.js
@@ -4,6 +4,12 @@
 import domReady from '@wordpress/dom-ready';
 import { dispatch, select } from '@wordpress/data';
 import { hasEventPastNotice } from './helpers/event';
+import { getBlockType, unregisterBlockType } from '@wordpress/blocks';
+
+/**
+ * Internal dependencies.
+ */
+import { getFromGlobal } from './helpers/globals';
 
 /**
  * Ensure Panels are Open for Events
@@ -48,4 +54,31 @@ domReady(() => {
 
 	// Display a notice for past events using the 'hasEventPastNotice' function.
 	hasEventPastNotice();
+});
+
+
+
+/**
+ * Remove Unwanted Blocks
+ *
+ * This script removes unwanted blocks from the localized array.
+ * It utilizes the `domReady` function to ensure the DOM is ready before execution.
+ * It iterates through the keys of the 'unregister_blocks' array obtained from the global scope,
+ * retrieves the block name, and unregisters the block using the `unregisterBlockType` function.
+ *
+ * @since 1.0.0
+ */
+
+// Execute the following code when the DOM is ready.
+domReady(() => {
+	// Iterate through keys of the 'unregister_blocks' array in the global scope.
+	Object.keys(getFromGlobal('misc.unregisterBlocks')).forEach((key) => {
+		// Retrieve the block name using the key.
+		const blockName = getFromGlobal('misc.unregisterBlocks')[key];
+
+		// Check if the block name is defined and unregister the block.
+		if (blockName && 'undefined' !== typeof getBlockType(blockName)) {
+			unregisterBlockType(blockName);
+		}
+	});
 });
