+++ webpack.config.js
@@ -12,7 +12,6 @@ module.exports = {
 	...defaultConfig,
 	entry: {
 		...defaultConfig.entry(),
-		admin: path.resolve(process.cwd(), 'src', 'admin.js'),
 		admin_style: path.resolve(process.cwd(), 'src', 'admin.scss'),
 		editor: path.resolve(process.cwd(), 'src', 'editor.js'),
 		panels: path.resolve(process.cwd(), 'src/panels', 'index.js'),
